ImportData now works but not with pre-existing databases

+++ Database/Model/Channel.cs
@@ -2,8 +2,8 @@
 {
     public class Channel
     {
-        public int ChannelId { get; set; }
-        public int GuildId { get; set; }
+        public ulong ChannelId { get; set; }
+        public ulong GuildId { get; set; }
         public string ChannelName { get; set; }
         public string CreationDate { get; set; }
     }
