+++ Database/Model/Role.cs
@@ -2,8 +2,8 @@
 {
     public class Role
     {
-        public int RoleId { get; set; }
-        public int GuildId { get; set; }
+        public ulong RoleId { get; set; }
+        public ulong GuildId { get; set; }
         public string RoleName { get; set; }
         public string RoleColor { get; set; }
         public string CreationDate { get; set; }
